alternative bash script
I had made a bash script as well as a python script, to call the API. I am adding this to my repo just for additional reference and completeness. It also serves as an alternative to the python script if necessary. Commands must be commented out for each curl however so it is not as controlled at the python script.